+++ @cybernaut/examples/chrome/vanilla.js
@@ -3,7 +3,7 @@
  */
 
 const {Chrome} = require('@cybernaut/chrome/lib/Chrome');
-const {Device} = require('@cybernaut/chrome/lib/Device');
+const {iPhone5} = require('@cybernaut/chrome/lib/MobileDevice');
 const {Engine} = require('@cybernaut/engine/lib/Engine');
 
 const {assert, perform} = new Engine();
@@ -12,7 +12,7 @@ const {assert, perform} = new Engine();
   const chrome = await Chrome.launchHeadless();
 
   try {
-    await perform(chrome.emulateDevice(Device.iPhone5()));
+    await perform(chrome.emulateMobileDevice(iPhone5()));
     await perform(chrome.navigateTo('https://www.example.com/'));
 
     await assert(chrome.pageTitle.is.equalTo('Example Domain'));